# packages/api/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/folders/{folderId}": {
      "get": {
        "operationId": "getFolder",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a folder by id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderGetResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Folders"
        ]
      }
    },
    "/folders": {
      "post": {
        "operationId": "createFolder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreateInputDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a folder.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderCreateResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Folders"
        ]
      }
    },
    "/folders/{folderId}/rescan": {
      "post": {
        "operationId": "rescanFolder",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Scan the underlying S3 location and update our local representation of it."
          }
        },
        "tags": [
          "Folders"
        ]
      }
    },
    "/folders/{folderId}/objects": {
      "get": {
        "operationId": "listFolderObjects",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List folder objects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderObjectsListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Folders"
        ]
      }
    },
...
  },
  "info": {
    "title": "@stellariscloud/api",
    "description": "The Stellaris Cloud core API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FolderGetResponse": {
        "type": "object",
        "properties": {
          "folder": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "ownerId": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "metadataLocation": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  },
                  "accessKeyId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "endpoint",
                  "region",
                  "bucket",
                  "accessKeyId"
                ]
              },
              "contentLocation": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  },
                  "accessKeyId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "endpoint",
                  "region",
                  "bucket",
                  "accessKeyId"
                ]
              }
            },
            "required": [
              "id",
              "ownerId",
              "name",
              "metadataLocation",
              "contentLocation"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "folder",
          "permissions"
        ]
      },
      "FolderCreateInputDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "metadataLocation": {
            "type": "object",
            "properties": {
              "serverLocationId": {
                "type": "string"
              },
              "userLocationId": {
                "type": "string"
              },
              "userLocationBucketOverride": {
                "type": "string"
              },
              "userLocationPrefixOverride": {
                "type": "string"
              },
              "accessKeyId": {
                "type": "string"
              },
              "secretAccessKey": {
                "type": "string"
              },
              "endpoint": {
                "type": "string"
              },
              "bucket": {
                "type": "string"
              },
              "region": {
                "type": "string"
              },
              "prefix": {
                "type": "string"
              }
            }
          },
          "contentLocation": {
            "type": "object",
            "properties": {
              "serverLocationId": {
                "type": "string"
              },
              "userLocationId": {
                "type": "string"
              },
              "userLocationBucketOverride": {
                "type": "string"
              },
              "userLocationPrefixOverride": {
                "type": "string"
              },
              "accessKeyId": {
                "type": "string"
              },
              "secretAccessKey": {
                "type": "string"
              },
              "endpoint": {
                "type": "string"
              },
              "bucket": {
                "type": "string"
              },
              "region": {
                "type": "string"
              },
              "prefix": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "name",
          "metadataLocation",
          "contentLocation"
        ]
      },
      "FolderCreateResponse": {
        "type": "object",
        "properties": {
          "folder": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "ownerId": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "metadataLocation": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  },
                  "accessKeyId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "endpoint",
                  "region",
                  "bucket",
                  "accessKeyId"
                ]
              },
              "contentLocation": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  },
                  "accessKeyId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "endpoint",
                  "region",
                  "bucket",
                  "accessKeyId"
                ]
              }
            },
            "required": [
              "id",
              "ownerId",
              "name",
              "metadataLocation",
              "contentLocation"
            ]
          }
        },
        "required": [
          "folder"
        ]
      },
      "FolderObjectsListResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "properties": {
              "totalCount": {
                "type": "number"
              }
            },
            "required": [
              "totalCount"
            ]
          },
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "objectKey": {
                  "type": "string"
                },
                "folderId": {
                  "type": "string"
                },
                "hash": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "lastModified": {
                  "type": "number"
                },
                "eTag": {
                  "type": "string"
                },
                "sizeBytes": {
                  "type": "number"
                },
                "mimeType": {
                  "type": "string"
                },
                "mediaType": {
                  "type": "string",
                  "enum": [
                    "IMAGE",
                    "VIDEO",
                    "AUDIO",
                    "DOCUMENT",
                    "UNKNOWN"
                  ]
                }
              },
              "required": [
                "id",
                "objectKey",
                "folderId",
                "lastModified",
                "eTag",
                "sizeBytes",
                "mimeType",
                "mediaType"
              ]
            }
          }
        },
        "required": [
          "meta",
          "result"
        ]
      },
...
    }
  }
}